* Add method for dragons steps processing
1) new internal method _next_step(row, col, previous)

 Part of logic of compute_intention_path responsible for defining
of the next cell between neighbor ones for the next move was replaced
to the above mentioned separate method.

2) new method dragon_move()

 dragon_move() is based on the part of replaced logic of end_turn
checking if there is a dragon on the hero cell and if it's there,
can the hero defeat him.

 But the initial logic has undergone changes. Namely now method defines
for every dragon next cell where to move it thanks to _next_pos() and if there is hero and
if it can defeat him.

 Results of conditions were left the same.

 If option dragon_moves is activated, thee method works as described above,
otherwise next position just rests the same and method execution is continued.